+++ client/src/components/Navigation/_NavItems/NavDropdown.vue
@@ -37,18 +37,6 @@
         />
       </slot>
     </template>
-    <b-dropdown-header
-      v-if="viewIsMobile"
-      class="for-mobile"
-    >
-      <span>{{ title }}</span>
-      <button
-        class="btn btn-link"
-        @click="hide()"
-      >
-        <i class="fas fa-times" />
-      </button>
-    </b-dropdown-header>
     <span
       class="content"
       :class="{