+++ client/src/views/partials/Navigation/Navigation.vue
@@ -1,7 +1,7 @@
 <template>
   <b-navbar
     toggleable="md"
-    :sticky="viewIsMD"
+    sticky
     class="navigation"
     :class="{
       'nav-not-visible': isLoggedIn,
@@ -19,6 +19,8 @@
       <b-collapse
         id="nav-collapse"
         is-nav
+        @show="disablePageScroll"
+        @hide="enablePageScroll"
       >
         <SideNavLoggedIn v-if="isLoggedIn" />
         <SideNavLoggedOut v-else />
@@ -120,6 +122,14 @@ export default {
       await DataBells.mutations.fetch()
       await DataConversations.initConversations()
     }
+  },
+  methods: {
+    enablePageScroll () {
+      document.body.classList.remove('nav-open')
+    },
+    disablePageScroll () {
+      document.body.classList.add('nav-open')
+    },
   },
 }
 </script>